Improve test architecture (#124)
* tests(Report): Setup an HTML reporter to improve result display

* tests(AdvancedDate): Increase delta to not fail on slow PC

* tests(...): Try to refactor how we manage login

* tests(...): Use utils function to test with authentication

* tests(.): Fix changes after last merge

* tests(.): Split tests into different directories

* .

* tests(thesis): split unit tests

* tests(thesis): reformat code according to IDE

* tests(thesis): Fix after reformat

* tests(thesis): Change spy by real mock

* tests(thesis): Refactor methods to search emails

* tests(thesis): Refactor before/after hook

* tests(thesis): Split last file
