+++ .gitignore
@@ -6,3 +6,4 @@
 **/.idea
 .DS_Store
 
+backend/reports
