+++ backend/package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "test": "jest --detectOpenHandles",
     "test:coverage": "jest --coverage --colors",
-    "test:unit": "npx jest --coverage --coverageDirectory=coverage/coverage_unit --colors --runInBand .*.unit.test.js",
-    "test:integration": "npx jest --coverage --coverageDirectory=coverage/coverage_integration  --runInBand --colors .*.integration.test.js",
+    "test:unit": "npx jest --coverage --coverageDirectory=coverage/coverage_unit --colors --runInBand units/*",
+    "test:integration": "npx jest --coverage --coverageDirectory=coverage/coverage_integration  --runInBand --colors integration/*",
     "start": "node index",
     "dev": "nodemon index"
   },
@@ -22,26 +22,37 @@
     "express": "^4.18.2",
     "express-session": "1.17.3",
     "formidable": "^3.5.1",
-    "fs": "^0.0.1-security",
-    "fs-extra": "^11.2.0",
-    "imap": "^0.8.19",
-    "jsonwebtoken": "^9.0.2",
     "morgan": "^1.10.0",
     "nodemailer": "^6.9.7",
     "passport": "^0.6.0",
-    "passport-local": "^1.0.0",
     "passport-saml": "^3.2.4",
-    "path": "^0.12.7",
-    "supertest": "^6.3.3",
     "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@types/imap": "^0.8.40",
     "@types/jest": "^29.5.6",
     "@types/supertest": "^2.0.16",
+    "fs-extra": "^11.2.0",
+    "imap": "^0.8.19",
     "jest": "^29.7.0",
-    "nodemon": "^3.0.1"
+    "jest-html-reporters": "^3.1.5",
+    "nodemon": "^3.0.1",
+    "supertest": "^6.3.3"
   },
   "jest": {
-    "collectCoverage": true
+    "collectCoverage": true,
+    "reporters": [
+      "default",
+      [
+        "jest-html-reporters",
+        {
+          "enableMergeData": true,
+          "dataMergeLevel": 1,
+          "openReport": true,
+          "publicPath": "./reports",
+          "filename": "index.html"
+        }
+      ]
+    ]
   }
 }
