+++ backend/test/integration/thesis.teacher.integration.test.js
@@ -1,42 +1,30 @@
 require('jest');
 // [i] This line setup the test database + load the environment variables. DON'T (RE)MOVE IT
-const { resetTestDatabase } = require('./integration_config');
+const { resetTestDatabase, initImapClient, closeImapClient, searchEmails} = require('../integration_config');
 
 const request = require("supertest");
-const { app } = require("../app");
-const thesisDao = require('../thesis_dao');
-const usersDao = require('../users_dao');
-const db = require('../db');
+const {app} = require("../../app");
+const utils = require("../utils");
+const thesisDao = require('../../thesis_dao');
+const usersDao = require('../../users_dao');
+const db = require('../../db');
 const path = require('path');
 const fs = require('fs');
 const fse = require('fs-extra');
-const imap = require("imap");
 
-// Mock Passport-SAML authenticate method
-jest.mock('passport-saml', () => {
-    const Strategy = jest.requireActual('passport-saml').Strategy;
-    return {
-      Strategy: class MockStrategy extends Strategy {
-        authenticate(req, options) {
-          const user = {
-            id: 'd279620',
-            name: 'Marco Rossi',
-            roles: ['teacher'], 
-          };
-          this.success(user);
-        }
-      },
-    };
+let agent;
+beforeAll(async () => {
+    await initImapClient();
+    agent = await utils.getMarcoRossiAgent(app);
 });
-
 beforeEach(() => {
     // Be sure that we are using a full clean database before each test
     resetTestDatabase();
     jest.restoreAllMocks();
 });
-beforeAll(async () => {
-    agent = request.agent(app);
-    await agent.get('/login');
+
+afterAll(async () => {
+    await closeImapClient();
 });
 
 describe('GET /api/teachers', () => {
@@ -83,9 +71,9 @@ describe('GET /api/externalCoSupervisors', () => {
         expect(response.status).toBe(200);
         expect(response.body).toEqual({
             externalCoSupervisors: [
-                { id: 1, surname: 'Amato', name: 'Alice', email: 'alice.amato@email.com' },
-                { id: 2, surname: 'Bianchi', name: 'Benjamin', email: 'benjamin.bianchi@email.com' },
-                { id: 3, surname: 'Colombo', name: 'Chiara', email: 'chiara.colombo@email.com' }
+                {id: 1, surname: 'Amato', name: 'Alice', email: 'alice.amato@email.com'},
+                {id: 2, surname: 'Bianchi', name: 'Benjamin', email: 'benjamin.bianchi@email.com'},
+                {id: 3, surname: 'Colombo', name: 'Chiara', email: 'chiara.colombo@email.com'}
             ]
         });
     });
@@ -148,7 +136,7 @@ describe('POST /api/teacher/thesis_proposals', () => {
     test('should create a new thesis proposal', async () => {
         const requestBody = {
             title: 'Test Thesis',
-            internal_co_supervisors_id: [], 
+            internal_co_supervisors_id: [],
             external_co_supervisors_id: [1],
             type: 'Bachelor',
             description: 'Test description',
@@ -250,11 +238,11 @@ describe('POST /api/teacher/thesis_proposals', () => {
         };
 
         const response = await agent
-          .post('/api/teacher/thesis_proposals')
-          .set('Accept', 'application/json')
-          .set('credentials', 'include')
-          .send(requestBody)
-          .expect(500);
+            .post('/api/teacher/thesis_proposals')
+            .set('Accept', 'application/json')
+            .set('credentials', 'include')
+            .send(requestBody)
+            .expect(500);
 
         // Expecting a 500 status code
         expect(response.status).toBe(500);
@@ -279,11 +267,11 @@ describe('POST /api/teacher/thesis_proposals', () => {
         };
 
         const response = await agent
-          .post('/api/teacher/thesis_proposals')
-          .set('Accept', 'application/json')
-          .set('credentials', 'include')
-          .send(requestBody)
-          .expect(500);
+            .post('/api/teacher/thesis_proposals')
+            .set('Accept', 'application/json')
+            .set('credentials', 'include')
+            .send(requestBody)
+            .expect(500);
 
         // Expecting a 500 status code
         expect(response.status).toBe(500);
@@ -300,7 +288,7 @@ describe('GET /api/keywords', () => {
 
         // Assertions
         expect(response.status).toBe(200);
-        expect(response.body).toEqual({ keywords: ["AI", "web development", "research", "reactive API", "QUIC"] });
+        expect(response.body).toEqual({keywords: ["AI", "web development", "research", "reactive API", "QUIC"]});
     });
 
     // TODO: Add tests on authentication controls
@@ -329,8 +317,8 @@ describe('GET /api/degrees', () => {
         // Assertions
         expect(response.status).toBe(200);
         expect(response.body).toEqual([
-            { cod_degree: 'L-08', title_degree: 'Ingegneria Elettronica' },
-            { cod_degree: 'LM-31', title_degree: 'Ingegneria dell\'Automazione' }
+            {cod_degree: 'L-08', title_degree: 'Ingegneria Elettronica'},
+            {cod_degree: 'LM-31', title_degree: 'Ingegneria dell\'Automazione'}
         ]);
     });
 
@@ -339,8 +327,8 @@ describe('GET /api/degrees', () => {
 
         // Send a request to the endpoint
         const response = await agent
-                        .get('/api/degrees')
-                        .set('credentials', 'include');
+            .get('/api/degrees')
+            .set('credentials', 'include');
 
         // Assertions
         expect(response.status).toBe(500);
@@ -377,8 +365,8 @@ describe('GET /api/thesis-proposals (teacher)', () => {
 
         // Send a request to the endpoint
         const response = await agent
-                               .get('/api/thesis-proposals')
-                               .set('credentials', 'include');
+            .get('/api/thesis-proposals')
+            .set('credentials', 'include');
 
         // Assertions
         expect(response.status).toBe(500);
@@ -410,10 +398,10 @@ describe('GET /api/thesis-proposals/:id (teacher)', () => {
             coSupervisors: {
                 internal: [],
                 external: [
-                    { 
+                    {
                         id: 1,
-                        name: 'Alice', 
-                        surname: 'Amato', 
+                        name: 'Alice',
+                        surname: 'Amato',
                         email: 'alice.amato@email.com',
                         co_supervisor_id: "1",
                         proposal_id: 1
@@ -427,7 +415,7 @@ describe('GET /api/thesis-proposals/:id (teacher)', () => {
             creation_date: '2023-10-10T10:45:50.121Z',
             expiration: '2024-11-10T23:59:59.999Z',
             level: 'LM',
-            cds: [{ cod_degree: 'L-08', title_degree: 'Ingegneria Elettronica'}],
+            cds: [{cod_degree: 'L-08', title_degree: 'Ingegneria Elettronica'}],
             keywords: ['AI', 'research', 'web development'],
             groups: ['Group1']
         });
@@ -515,7 +503,7 @@ describe('PUT /api/thesis-proposals/:id', () => {
             creation_date: '2023-10-10T10:45:50.121Z',
             expiration: '2025-11-10T23:59:59.999Z',
             level: 'LM',
-            cds: [{ cod_degree: 'L-08', title_degree: 'Ingegneria Elettronica'}],
+            cds: [{cod_degree: 'L-08', title_degree: 'Ingegneria Elettronica'}],
             keywords: ['AI', 'another keyword', 'research', 'web development'],
             groups: ['Group1']
         });
@@ -526,11 +514,11 @@ describe('PUT /api/thesis-proposals/:id', () => {
         const response = await agent
             .put(`/api/thesis-proposals/3`)
             .set('credentials', 'include')
-            .send({ });
+            .send({});
 
         // Assert the response
         expect(response.status).toBe(403);
-        expect(response.body).toEqual({ message: 'Cannot edit a proposal with accepted applications.' });
+        expect(response.body).toEqual({message: 'Cannot edit a proposal with accepted applications.'});
     });
 
     test('should return 404 error if the proposal does not exist', async () => {
@@ -556,7 +544,7 @@ describe('PUT /api/thesis-proposals/:id', () => {
 
         // Assert the response
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: `Thesis proposal with id 32 not found.` });
+        expect(response.body).toEqual({message: `Thesis proposal with id 32 not found.`});
     });
 
     test('should return error 400 if some properties are missing', async () => {
@@ -585,18 +573,85 @@ describe('PUT /api/thesis-proposals/:id', () => {
         expect(response.body).toEqual(
             {
                 "errors":
-                [
-                    {
-                        "code": "invalid_type",
-                        "expected": "array",
-                        "message": "Required",
-                        "path": ["cds"],
-                        "received": "undefined"
-                    }
-                ],
+                    [
+                        {
+                            "code": "invalid_type",
+                            "expected": "array",
+                            "message": "Required",
+                            "path": ["cds"],
+                            "received": "undefined"
+                        }
+                    ],
                 "message": "Some properties are missing or invalid."
             }
         );
+    });
+    test('should return 400 if expiration date is invalid', async () => {
+        const responseGet = await agent
+            .get('/api/thesis-proposals/1')
+            .set('Accept', 'application/json')
+            .set('credentials', 'include')
+            .send();
+        expect(responseGet.status).toBe(200);
+
+        const body = responseGet.body;
+        const updatedBody = {
+            title: body.title + ' updated',
+            type: body.type + ' updated',
+            description: body.description + ' updated',
+            required_knowledge: body.requiredKnowledge + ' updated',
+            notes: body.notes + ' updated',
+            expiration: 'something-invalid',
+            level: body.level,
+            cds: ['L-08'],
+            keywords: [...body.keywords, 'another keyword'],
+            internal_co_supervisors_id: [],
+            external_co_supervisors_id: []
+        };
+        const response = await agent
+            .put(`/api/thesis-proposals/1`)
+            .set('credentials', 'include')
+            .send(updatedBody);
+
+        expect(response.status).toBe(400);
+        expect(response.body.message).toEqual("Some properties are missing or invalid.");
+        expect(response.body).toHaveProperty('errors');
+        expect(response.body.errors).toHaveLength(1);
+        expect(response.body.errors[0].code).toEqual('invalid_date');
+    });
+
+    test('should return 400 if expiration date is in the past', async () => {
+        const responseGet = await agent
+            .get('/api/thesis-proposals/1')
+            .set('Accept', 'application/json')
+            .set('credentials', 'include')
+            .send();
+        expect(responseGet.status).toBe(200);
+
+        const body = responseGet.body;
+        const updatedBody = {
+            title: body.title + ' updated',
+            type: body.type + ' updated',
+            description: body.description + ' updated',
+            required_knowledge: body.requiredKnowledge + ' updated',
+            notes: body.notes + ' updated',
+            expiration: '2000-11-10',
+            level: body.level,
+            cds: ['L-08'],
+            keywords: [...body.keywords, 'another keyword'],
+            internal_co_supervisors_id: [],
+            external_co_supervisors_id: []
+        };
+        const response = await agent
+            .put(`/api/thesis-proposals/1`)
+            .set('credentials', 'include')
+            .send(updatedBody);
+
+        expect(response.status).toBe(400);
+        expect(response.body.message).toEqual("Some properties are missing or invalid.");
+        expect(response.body).toHaveProperty('errors');
+        expect(response.body.errors).toHaveLength(1);
+        expect(response.body.errors[0].code).toEqual('invalid_date');
     });
 
     test('should return 500 error', async () => {
@@ -652,7 +707,7 @@ describe('GET /api/teacher/applications/:proposal_id', () => {
         // Assertions
         expect(response.status).toBe(200);
         expect(response.body).toEqual([
-            { application_id: 2, id: 's321529', name: 'Matteo', surname: 'Ladrat', status: 'waiting for approval' }
+            {application_id: 2, id: 's321529', name: 'Matteo', surname: 'Ladrat', status: 'waiting for approval'}
         ]);
     });
 
@@ -681,58 +736,25 @@ describe('PATCH /api/teacher/applications/accept/:proposal_id', () => {
 
         // Act
         const response = await agent
-          .patch("/api/teacher/applications/accept/2")
-          .send({ student_id: 's318952' })
-          .set('credentials', 'include')
-          .expect(200);
+            .patch("/api/teacher/applications/accept/2")
+            .send({student_id: 's318952'})
+            .set('credentials', 'include')
+            .expect(200);
 
         // Assert
-        expect(response.body).toEqual({ message: 'Thesis accepted and others rejected successfully' });
+        expect(response.body).toEqual({message: 'Thesis accepted and others rejected successfully'});
 
         // Wait for a moment to allow the email to be processed (adjust the timing as needed)
         await new Promise((resolve) => setTimeout(resolve, 5000));
 
-        // Set up IMAP client
-        const imapClient = new imap({
-        user: process.env.TM_SMTP_USERNAME, 
-        password: process.env.TM_SMTP_PASSWORD,
-        host: "imap.ethereal.email", // IMAP server host
-        port: 993, // IMAP server port
-        tls: true,
-        tlsOptions: { rejectUnauthorized: false }, 
-        });
-
-        // Connect to the IMAP server
-        await new Promise((resolve, reject) => {
-        imapClient.once("ready", resolve);
-        imapClient.once("error", reject);
-        imapClient.connect();
-        });
-
-        // Open the Inbox
-        const openInbox = await new Promise((resolve, reject) => {
-        imapClient.openBox("INBOX", true, (err, box) => {
-            if (err) reject(err);
-            resolve(box);
-        });
-        });
-
         const subject = 'Application status changed - PERFORMANCE EVALUATION OF KAFKA CLIENTS USING A REACTIVE API';
+        const to = 's318952@studenti.polito.it';
         // Search for the email (adjust criteria as needed)
-        const searchResults = await new Promise((resolve, reject) => {
-            imapClient.search(['UNSEEN', ['TO', 's318952@studenti.polito.it'], ['SUBJECT', subject]], (err, results) => {
-            if (err) reject(err);
-            resolve(results);
-            });
-        });
+        const searchResults = await searchEmails(to, subject);
 
         // Assert that the email has been received
         expect(searchResults.length).toBeGreaterThan(0);
-
-        // Close the IMAP connection
-        imapClient.end();
-
-    },10000);
+    }, 10000);
     test('should return 400 error if is missing student_id', async () => {
         // Act
         const response = await agent
@@ -741,25 +763,25 @@ describe('PATCH /api/teacher/applications/accept/:proposal_id', () => {
 
         // Assert
         expect(response.status).toBe(400);
-        expect(response.body).toEqual({ message: 'Missing required fields.' });
+        expect(response.body).toEqual({message: 'Missing required fields.'});
     });
     test('should refuse if the thesis doesn\'t exist', async () => {
         // Act
         const response = await agent
             .patch("/api/teacher/applications/accept/5")
-            .send({ student_id: 's318952' })
+            .send({student_id: 's318952'})
             .set('credentials', 'include');
 
         // Assert
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: 'Thesis proposal with id 5 not found, cannot accept this application.' });
+        expect(response.body).toEqual({message: 'Thesis proposal with id 5 not found, cannot accept this application.'});
     });
     test('should return 404 error if no application has been found', async () => {
         // Act
         const response = await agent
             .patch("/api/teacher/applications/accept/2")
             .set('credentials', 'include')
-            .send({ student_id: 's4' });
+            .send({student_id: 's4'});
 
         // Assert
         expect(response.status).toBe(404);
@@ -774,9 +796,9 @@ describe('PATCH /api/teacher/applications/accept/:proposal_id', () => {
 
         // Act
         const response = await agent
-          .patch(`/api/teacher/applications/accept/${proposalId}`)
-          .set('credentials', 'include')
-          .send({ student_id: studentId });
+            .patch(`/api/teacher/applications/accept/${proposalId}`)
+            .set('credentials', 'include')
+            .send({student_id: studentId});
 
         // Assert
         expect(response.status).toBe(500);
@@ -794,56 +816,24 @@ describe('PATCH /api/teacher/applications/reject/:proposal_id', () => {
         // Act
         const response = await agent
             .patch("/api/teacher/applications/reject/2")
-            .send({ student_id: 's321529' })
+            .send({student_id: 's321529'})
             .set('credentials', 'include')
             .expect(200);
 
         // Assert
-        expect(response.body).toEqual({ message: 'Thesis successfully rejected' });
+        expect(response.body).toEqual({message: 'Thesis successfully rejected'});
 
         // Wait for a moment to allow the email to be processed (adjust the timing as needed)
         await new Promise((resolve) => setTimeout(resolve, 5000));
 
-        // Set up IMAP client
-        const imapClient = new imap({
-        user: process.env.TM_SMTP_USERNAME, 
-        password: process.env.TM_SMTP_PASSWORD,
-        host: "imap.ethereal.email", // IMAP server host
-        port: 993, // IMAP server port
-        tls: true,
-        tlsOptions: { rejectUnauthorized: false }, 
-        });
-
-        // Connect to the IMAP server
-        await new Promise((resolve, reject) => {
-        imapClient.once("ready", resolve);
-        imapClient.once("error", reject);
-        imapClient.connect();
-        });
-
-        // Open the Inbox
-        const openInbox = await new Promise((resolve, reject) => {
-        imapClient.openBox("INBOX", true, (err, box) => {
-            if (err) reject(err);
-            resolve(box);
-        });
-        });
-
         const subject = 'Application status changed - PERFORMANCE EVALUATION OF KAFKA CLIENTS USING A REACTIVE API';
+        const to = '321529@studenti.polito.it';
         // Search for the email (adjust criteria as needed)
-        const searchResults = await new Promise((resolve, reject) => {
-            imapClient.search(['UNSEEN', ['TO', '321529@studenti.polito.it'], ['SUBJECT', subject]], (err, results) => {
-            if (err) reject(err);
-            resolve(results);
-            });
-        });
+        const searchResults = await searchEmails(to, subject);
 
         // Assert that the email has been received
         expect(searchResults.length).toBeGreaterThan(0);
-
-        // Close the IMAP connection
-        imapClient.end()
-    },10000);
+    }, 10000);
     test('should return 400 error if is missing student_id', async () => {
         // Act
         const response = await agent
@@ -852,25 +842,25 @@ describe('PATCH /api/teacher/applications/reject/:proposal_id', () => {
 
         // Assert
         expect(response.status).toBe(400);
-        expect(response.body).toEqual({ message: 'Missing required fields.' });
+        expect(response.body).toEqual({message: 'Missing required fields.'});
     });
     test('should refuse if the thesis doesn\'t exist', async () => {
         // Act
         const response = await agent
             .patch("/api/teacher/applications/reject/5")
-            .send({ student_id: 's318952' })
+            .send({student_id: 's318952'})
             .set('credentials', 'include');
 
         // Assert
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: 'Thesis proposal with id 5 not found, cannot reject this application.' });
+        expect(response.body).toEqual({message: 'Thesis proposal with id 5 not found, cannot reject this application.'});
     });
     test('should return 404 error if no application has been found', async () => {
         // Act
         const response = await agent
             .patch("/api/teacher/applications/reject/2")
             .set('credentials', 'include')
-            .send({ student_id: 's4' });
+            .send({student_id: 's4'});
 
         // Assert
         expect(response.status).toBe(404);
@@ -887,7 +877,7 @@ describe('PATCH /api/teacher/applications/reject/:proposal_id', () => {
         const response = await agent
             .patch(`/api/teacher/applications/reject/${proposalId}`)
             .set('credentials', 'include')
-            .send({ student_id: studentId });
+            .send({student_id: studentId});
 
         // Assert
         expect(response.status).toBe(500);
@@ -948,7 +938,7 @@ describe('GET /api/student/:id/career', () => {
         expect(response.status).toBe(404);
     });
     test('should return 500 if internal server error happens', async () => {
-        
+
         jest.spyOn(usersDao, 'getStudentById').mockRejectedValueOnce(new Error());
         const response = await agent
             .get('/api/student/s318952/career')
@@ -975,7 +965,7 @@ describe('DELETE /api/thesis-proposals/:id', () => {
     test('should delete a thesis proposal and cancel pending applications', async () => {
         db.prepare('INSERT INTO thesisApplication (student_id, proposal_id, creation_date, status) VALUES (?, ?, ?, ?)')
             .run('s318952', 1, new Date().toISOString(), 'waiting for approval');
-        
+
         const id = 1;
 
         // Make a request to the endpoint
@@ -986,50 +976,17 @@ describe('DELETE /api/thesis-proposals/:id', () => {
         // Assertions
         expect(response.status).toBe(204);
 
-         // Wait for a moment to allow the email to be processed (adjust the timing as needed)
-         await new Promise((resolve) => setTimeout(resolve, 5000));
+        // Wait for a moment to allow the email to be processed (adjust the timing as needed)
+        await new Promise((resolve) => setTimeout(resolve, 5000));
 
-         // Set up IMAP client
-         const imapClient = new imap({
-         user: process.env.TM_SMTP_USERNAME, 
-         password: process.env.TM_SMTP_PASSWORD,
-         host: "imap.ethereal.email", // IMAP server host
-         port: 993, // IMAP server port
-         tls: true,
-         tlsOptions: { rejectUnauthorized: false }, 
-         });
- 
-         // Connect to the IMAP server
-         await new Promise((resolve, reject) => {
-         imapClient.once("ready", resolve);
-         imapClient.once("error", reject);
-         imapClient.connect();
-         });
- 
-         // Open the Inbox
-         const openInbox = await new Promise((resolve, reject) => {
-         imapClient.openBox("INBOX", true, (err, box) => {
-             if (err) reject(err);
-             resolve(box);
-         });
-         });
- 
-         const subject = 'Application status changed - PERFORMANCE EVALUATION OF KAFKA CLIENTS USING A REACTIVE API';
-         // Search for the email (adjust criteria as needed)
-         const searchResults = await new Promise((resolve, reject) => {
-             imapClient.search(['UNSEEN', ['TO', 's318952@studenti.polito.it'], ['SUBJECT', subject]], (err, results) => {
-             if (err) reject(err);
-             resolve(results);
-             });
-         });
- 
-         // Assert that the email has been received
-         expect(searchResults.length).toBeGreaterThan(0);
- 
-         // Close the IMAP connection
-         imapClient.end();
- 
-    },10000);
+        const subject = 'Application status changed - PERFORMANCE EVALUATION OF KAFKA CLIENTS USING A REACTIVE API';
+        const to = 's318952@studenti.polito.it';
+        // Search for the email (adjust criteria as needed)
+        const searchResults = await searchEmails(to, subject);
+
+        // Assert that the email has been received
+        expect(searchResults.length).toBeGreaterThan(0);
+    }, 10000);
     test('should refuse to delete a thesis proposal with accepted application', async () => {
         const id = 3;
 
@@ -1040,7 +997,7 @@ describe('DELETE /api/thesis-proposals/:id', () => {
 
         // Assertions
         expect(response.status).toBe(403);
-        expect(response.body).toEqual({ message: 'Some applications has been accepted and, therefore, you can\'t delete this thesis' });
+        expect(response.body).toEqual({message: 'Some applications has been accepted and, therefore, you can\'t delete this thesis'});
     });
     test('should refuse to delete a inexistent thesis proposal', async () => {
         const id = 4;
@@ -1052,13 +1009,13 @@ describe('DELETE /api/thesis-proposals/:id', () => {
 
         // Assertions
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: 'No thesis proposal with id 4 found' });
+        expect(response.body).toEqual({message: 'No thesis proposal with id 4 found'});
     });
     test('should refuse to delete a inexistent thesis proposal (because created in the future)', async () => {
         db.prepare('INSERT INTO thesisProposal (proposal_id, title, supervisor_id, type, description, required_knowledge, notes, creation_date, expiration, level, is_deleted, is_archived)  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)')
-        .run(4, 'Title', 'd279620', 'research project', 'Description', 'Required knowledge', 'Notes', '2030-10-10T10:45:50.121Z', '2032-11-10T23:59:59.999Z', 'LM', 0, 0);
+            .run(4, 'Title', 'd279620', 'research project', 'Description', 'Required knowledge', 'Notes', '2030-10-10T10:45:50.121Z', '2032-11-10T23:59:59.999Z', 'LM', 0, 0);
         db.prepare('INSERT INTO proposalCds (proposal_id, cod_degree) VALUES (?, ?)')
-        .run(4, 'L-08');
+            .run(4, 'L-08');
 
         // Make a request to the endpoint
         const response = await agent
@@ -1067,13 +1024,13 @@ describe('DELETE /api/thesis-proposals/:id', () => {
 
         // Assertions
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: 'No thesis proposal with id 4 found' });
+        expect(response.body).toEqual({message: 'No thesis proposal with id 4 found'});
     });
     test('should refuse to delete an expired thesis proposal', async () => {
         db.prepare('INSERT INTO thesisProposal (proposal_id, title, supervisor_id, type, description, required_knowledge, notes, creation_date, expiration, level)  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)')
-        .run(4, 'Title', 'd279620', 'research project', 'Description', 'Required knowledge', 'Notes', '2020-10-10T10:45:50.121Z', '2022-11-10T23:59:59.999Z', 'LM');
+            .run(4, 'Title', 'd279620', 'research project', 'Description', 'Required knowledge', 'Notes', '2020-10-10T10:45:50.121Z', '2022-11-10T23:59:59.999Z', 'LM');
         db.prepare('INSERT INTO proposalCds (proposal_id, cod_degree) VALUES (?, ?)')
-        .run(4, 'L-08');
+            .run(4, 'L-08');
         // Make a request to the endpoint
         const response = await agent
             .delete(`/api/thesis-proposals/4`)
@@ -1081,13 +1038,13 @@ describe('DELETE /api/thesis-proposals/:id', () => {
 
         // Assertions
         expect(response.status).toBe(403);
-        expect(response.body).toEqual({ message: 'You can\'t delete a thesis already expired' });
+        expect(response.body).toEqual({message: 'You can\'t delete a thesis already expired'});
     });
     test('should refuse to delete a thesis for which the teacher is not the supervisor', async () => {
         db.prepare('INSERT INTO thesisProposal (proposal_id, title, supervisor_id, type, description, required_knowledge, notes, creation_date, expiration, level)  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)')
-        .run(4, 'Title', 'd370335', 'research project', 'Description', 'Required knowledge', 'Notes', '2020-10-10T10:45:50.121Z', '2032-11-10T23:59:59.999Z', 'LM');
+            .run(4, 'Title', 'd370335', 'research project', 'Description', 'Required knowledge', 'Notes', '2020-10-10T10:45:50.121Z', '2032-11-10T23:59:59.999Z', 'LM');
         db.prepare('INSERT INTO proposalCds (proposal_id, cod_degree) VALUES (?, ?)')
-        .run(4, 'L-08');
+            .run(4, 'L-08');
         // Make a request to the endpoint
         const response = await agent
             .delete(`/api/thesis-proposals/4`)
@@ -1095,7 +1052,7 @@ describe('DELETE /api/thesis-proposals/:id', () => {
 
         // Assertions
         expect(response.status).toBe(403);
-        expect(response.body).toEqual({ message: 'You are not the supervisor of this thesis' });
+        expect(response.body).toEqual({message: 'You are not the supervisor of this thesis'});
     });
     test('should handle errors and return the appropriate response', async () => {
         const id = 'nonExistingProposalId';
@@ -1107,7 +1064,7 @@ describe('DELETE /api/thesis-proposals/:id', () => {
 
         // Assertions
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: `No thesis proposal with id ${id} found` });
+        expect(response.body).toEqual({message: `No thesis proposal with id ${id} found`});
     });
     test('should handle unexpected errors and return 500 Internal Server Error', async () => {
         jest.spyOn(thesisDao, 'deleteThesisProposalById').mockRejectedValueOnce(new Error());
@@ -1119,7 +1076,7 @@ describe('DELETE /api/thesis-proposals/:id', () => {
 
         // Assertions
         expect(response.status).toBe(500);
-        expect(response.body).toEqual({ message: 'Internal Server Error' });
+        expect(response.body).toEqual({message: 'Internal Server Error'});
     });
 });
 
@@ -1158,7 +1115,7 @@ describe('PATCH /api/thesis-proposals/archive/:id', () => {
 
         // Assertions
         expect(response.status).toBe(403);
-        expect(response.body).toEqual({ message: 'Some applications has been accepted and, therefore, you can\'t archive this thesis'});
+        expect(response.body).toEqual({message: 'Some applications has been accepted and, therefore, you can\'t archive this thesis'});
     });
     test('should not archive a inexistent thesis proposal', async () => {
         const id = 4;
@@ -1170,14 +1127,14 @@ describe('PATCH /api/thesis-proposals/archive/:id', () => {
 
         // Assertions
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: 'No thesis proposal with id 4 found'});
+        expect(response.body).toEqual({message: 'No thesis proposal with id 4 found'});
     });
     test('should not archive a thesis proposal created in the future', async () => {
-        
+
         db.prepare('INSERT INTO thesisProposal (proposal_id, title, supervisor_id, type, description, required_knowledge, notes, creation_date, expiration, level)  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)')
-        .run(4, 'Title', 'd279620', 'research project', 'Description', 'Required knowledge', 'Notes', '2030-10-10T10:45:50.121Z', '2032-11-10T23:59:59.999Z', 'LM');
+            .run(4, 'Title', 'd279620', 'research project', 'Description', 'Required knowledge', 'Notes', '2030-10-10T10:45:50.121Z', '2032-11-10T23:59:59.999Z', 'LM');
         db.prepare('INSERT INTO proposalCds (proposal_id, cod_degree) VALUES (?, ?)')
-        .run(4, 'L-08');
+            .run(4, 'L-08');
 
         const id = 4;
 
@@ -1188,14 +1145,14 @@ describe('PATCH /api/thesis-proposals/archive/:id', () => {
 
         // Assertions
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: 'No thesis proposal with id 4 found'});
+        expect(response.body).toEqual({message: 'No thesis proposal with id 4 found'});
     });
     test('should not archive a thesis proposal already expired (and archived)', async () => {
-        
+
         db.prepare('INSERT INTO thesisProposal (proposal_id, title, supervisor_id, type, description, required_knowledge, notes, creation_date, expiration, level)  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)')
-        .run(4, 'Title', 'd279620', 'research project', 'Description', 'Required knowledge', 'Notes', '2020-10-10T10:45:50.121Z', '2022-11-10T23:59:59.999Z', 'LM');
+            .run(4, 'Title', 'd279620', 'research project', 'Description', 'Required knowledge', 'Notes', '2020-10-10T10:45:50.121Z', '2022-11-10T23:59:59.999Z', 'LM');
         db.prepare('INSERT INTO proposalCds (proposal_id, cod_degree) VALUES (?, ?)')
-        .run(4, 'L-08');
+            .run(4, 'L-08');
 
         const id = 4;
 
@@ -1206,14 +1163,14 @@ describe('PATCH /api/thesis-proposals/archive/:id', () => {
 
         // Assertions
         expect(response.status).toBe(403);
-        expect(response.body).toEqual({ message: 'You can\'t archive a thesis already expired'});
+        expect(response.body).toEqual({message: 'You can\'t archive a thesis already expired'});
     });
     test('should not archive a thesis proposal if you are not the supervisor', async () => {
-        
+
         db.prepare('INSERT INTO thesisProposal (proposal_id, title, supervisor_id, type, description, required_knowledge, notes, creation_date, expiration, level)  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)')
-        .run(4, 'Title', 'd370335', 'research project', 'Description', 'Required knowledge', 'Notes', '2022-10-10T10:45:50.121Z', '2025-11-10T23:59:59.999Z', 'LM');
+            .run(4, 'Title', 'd370335', 'research project', 'Description', 'Required knowledge', 'Notes', '2022-10-10T10:45:50.121Z', '2025-11-10T23:59:59.999Z', 'LM');
         db.prepare('INSERT INTO proposalCds (proposal_id, cod_degree) VALUES (?, ?)')
-        .run(4, 'L-08');
+            .run(4, 'L-08');
 
         const id = 4;
 
@@ -1224,7 +1181,7 @@ describe('PATCH /api/thesis-proposals/archive/:id', () => {
 
         // Assertions
         expect(response.status).toBe(403);
-        expect(response.body).toEqual({ message: 'You are not the supervisor of this thesis'});
+        expect(response.body).toEqual({message: 'You are not the supervisor of this thesis'});
     });
     test('should handle errors and return the appropriate response', async () => {
         const id = 'nonExistingProposalId';
@@ -1236,7 +1193,7 @@ describe('PATCH /api/thesis-proposals/archive/:id', () => {
 
         // Assertions
         expect(response.status).toBe(404);
-        expect(response.body).toEqual({ message: `No thesis proposal with id ${id} found` });
+        expect(response.body).toEqual({message: `No thesis proposal with id ${id} found`});
     });
     test('should handle unexpected errors and return 500 Internal Server Error', async () => {
         jest.spyOn(thesisDao, 'archiveThesisProposalById').mockRejectedValueOnce(new Error());
@@ -1248,7 +1205,7 @@ describe('PATCH /api/thesis-proposals/archive/:id', () => {
 
         // Assertions
         expect(response.status).toBe(500);
-        expect(response.body).toEqual({ message: 'Internal Server Error' });
+        expect(response.body).toEqual({message: 'Internal Server Error'});
     });
 });
 
@@ -1257,38 +1214,38 @@ describe('GET /api/teacher/uploads/:stud_id/:app_id', () => {
         // Use fs-extra's remove method to delete the directory and its contents
         await fse.remove('uploads');
     });
-    
+
     test('returns 404 for non-existing student', async () => {
-      const nonExistingStudentId = 10000;
-      const nonExistingApplicationId = 10000;
-  
-      // Make the request
-      const response = await agent
-        .get(`/api/teacher/uploads/${nonExistingStudentId}/${nonExistingApplicationId}`)
-        .set('credentials', 'include')
-        .expect(404);
-  
-      // Assert the response
-      expect(response.body).toEqual({message : 'Student with id 10000 not found.'});
+        const nonExistingStudentId = 10000;
+        const nonExistingApplicationId = 10000;
+
+        // Make the request
+        const response = await agent
+            .get(`/api/teacher/uploads/${nonExistingStudentId}/${nonExistingApplicationId}`)
+            .set('credentials', 'include')
+            .expect(404);
+
+        // Assert the response
+        expect(response.body).toEqual({message: 'Student with id 10000 not found.'});
     });
 
     test('returns 404 for non-existing application', async () => {
         const nonExistingStudentId = 's318952';
         const nonExistingApplicationId = 10000;
-    
+
         // Make the request
         const response = await agent
-          .get(`/api/teacher/uploads/${nonExistingStudentId}/${nonExistingApplicationId}`)
-          .set('credentials', 'include')
-          .expect(404);
-    
+            .get(`/api/teacher/uploads/${nonExistingStudentId}/${nonExistingApplicationId}`)
+            .set('credentials', 'include')
+            .expect(404);
+
         // Assert the response
-        expect(response.body).toEqual({message : 'Application with id 10000 not found.'});
+        expect(response.body).toEqual({message: 'Application with id 10000 not found.'});
     });
 
     test('handle unexpected errors and return status 500', async () => {
         db.prepare('INSERT INTO thesisApplication (student_id, proposal_id, creation_date, status) VALUES (?, ?, ?, ?)')
-        .run('s318952', 2, new Date().toISOString(), 'waiting for approval');
+            .run('s318952', 2, new Date().toISOString(), 'waiting for approval');
 
         const nonExistingStudentId = 's318952';
         const nonExistingApplicationId = 2;
@@ -1296,67 +1253,67 @@ describe('GET /api/teacher/uploads/:stud_id/:app_id', () => {
         jest.spyOn(usersDao, 'getStudentById').mockRejectedValueOnce(new Error());
         // Make the request
         const response = await agent
-          .get(`/api/teacher/uploads/${nonExistingStudentId}/${nonExistingApplicationId}`)
-          .set('credentials', 'include')
-          .expect(500);
-    
+            .get(`/api/teacher/uploads/${nonExistingStudentId}/${nonExistingApplicationId}`)
+            .set('credentials', 'include')
+            .expect(500);
+
         // Assert the response
         expect(response.body).toEqual('Internal Server Error');
     });
 
     test('should return an empty JSON object if directory doesn\'t exist', async () => {
         db.prepare('INSERT INTO thesisApplication (student_id, proposal_id, creation_date, status) VALUES (?, ?, ?, ?)')
-        .run('s318952', 2, new Date().toISOString(), 'waiting for approval');
+            .run('s318952', 2, new Date().toISOString(), 'waiting for approval');
 
-        const studentId = 's318952'; 
-        const applicationId = 2; 
-    
+        const studentId = 's318952';
+        const applicationId = 2;
+
         // Make the request
         const response = await agent
-          .get(`/api/teacher/uploads/${studentId}/${applicationId}`)
-          .set('credentials', 'include')
-          .expect(200);
-    
+            .get(`/api/teacher/uploads/${studentId}/${applicationId}`)
+            .set('credentials', 'include')
+            .expect(200);
+
         expect(response.body).toEqual({});
     });
 
     test('should return an empty JSON object if directory exists but is empty', async () => {
-        
+
         db.prepare('INSERT INTO thesisApplication (student_id, proposal_id, creation_date, status) VALUES (?, ?, ?, ?)')
-        .run('s318952', 2, new Date().toISOString(), 'waiting for approval');
-        
+            .run('s318952', 2, new Date().toISOString(), 'waiting for approval');
+
         const studentId = 's318952';
         const applicationId = 2;
-        const dir = path.join(__dirname, '../uploads', studentId, applicationId.toString());
-    
+        const dir = path.join(__dirname, '../../uploads', studentId, applicationId.toString());
+
         // Create the directory without any files
-        fs.mkdirSync(dir, { recursive: true });
-    
+        fs.mkdirSync(dir, {recursive: true});
+
         const response = await agent
-          .get(`/api/teacher/uploads/${studentId}/${applicationId}`)
-          .expect(200);
-    
+            .get(`/api/teacher/uploads/${studentId}/${applicationId}`)
+            .expect(200);
+
         expect(response.body).toEqual({});
     });
 
     test('should return the first file if directory exists and contains files', async () => {
         db.prepare('INSERT INTO thesisApplication (student_id, proposal_id, creation_date, status) VALUES (?, ?, ?, ?)')
-        .run('s318952', 2, new Date().toISOString(), 'waiting for approval');
+            .run('s318952', 2, new Date().toISOString(), 'waiting for approval');
 
-         // Create a directory with a sample PDF file
-        const dir = path.join(__dirname, '../uploads', 's318952', '2');
-        fs.mkdirSync(dir, { recursive: true });
+        // Create a directory with a sample PDF file
+        const dir = path.join(__dirname, '../../uploads', 's318952', '2');
+        fs.mkdirSync(dir, {recursive: true});
         fs.writeFileSync(path.join(dir, 'sample.pdf'), Buffer.alloc(0));
 
-        const studentId = 's318952'; 
-        const applicationId = 2; 
-    
+        const studentId = 's318952';
+        const applicationId = 2;
+
         // Make the request
         const response = await agent
-          .get(`/api/teacher/uploads/${studentId}/${applicationId}`)
-          .set('credentials', 'include')
-          .expect(200);
-    
+            .get(`/api/teacher/uploads/${studentId}/${applicationId}`)
+            .set('credentials', 'include')
+            .expect(200);
+
         expect(response.header['content-type']).toEqual('application/pdf');
     });
 });
